Get meta data in the current track

Get meta data in the current track

# S3 method for TRACK_META
$(x, name)

Arguments

x
Always use TRACK_META.

name
Name of the meta name. For all supported names, type names(TRACK_META).

Details

The variable TRACK_META can only be used to get meta data from the "current" track. If the current track is not the one you want, you can first use set_current_track to set the current track.

There are following meta data for the current track:

xlim:

Data range on x-axis.

xmin:

xlim[1].

xmax:

xlim[2].

xrange:

xlim[2] - xlim[1].

xcenter:

mean(xlim).

theta_lim:

Range of the angles on the spiral, measured in radians.

theta_min:

theta_lim[1].

theta_max:

theta_lim[2].

theta_range:

theta_lim[2] - theta_lim[1].

theta_center:

mean(theta_lim).

ylim:

Data range on y-axis.

ymin:

ylim[1].

ymax:

ylim[2].

yrange:

ylim[2] - ylim[1].

ycenter:

mean(ylim).

rel_height:

Fraction of height of the track to the distance between two neighbouring loops.

abs_height:

The height of the track, which is rel_height multiplied by the distance between two neighbouring loops.

track_index:

Current track index.

Examples

# There is no example
NULL

#> NULL