Draw horizon chart along the spiral

Draw horizon chart along the spiral

spiral_horizon(x, y, n_slices = 4, slice_size, pos_fill = "#D73027", neg_fill = "#313695",
    use_bars = FALSE, bar_width = min(diff(x)),
    negative_from_top = FALSE, track_index = current_track_index())

Arguments

x
X-locations of the data points.

y
Y-locations of the data points.

n_slices
Number of slices.

slice_size
Size of the slices. The final number of sizes is ceiling(max(abs(y))/slice_size).

pos_fill
Colors for positive values.

neg_fill
Colors for negative values.

use_bars
Whether to use bars?

bar_width
Width of bars.

negative_from_top
Should negative distribution be drawn from the top?

track_index
Index of the track.

Details

Since the track height is very small in the spiral, horizon chart visualization is a efficient way to visualize distribution-like graphics.

Value

A list of three objects:

a color mapping function for positive colors.

a color mapping function for negative colors.

a vector of intervals that split the data.

Examples

# \donttest{
# df = cranlogs::cran_downloads("ggplot2", from="2014-01-01")
# the data frame `df` is already downloaded
df = readRDS(system.file("extdata", "ggplot2_download.rds", package = "spiralize"))

#> Warning: cannot open compressed file '', probable reason 'No such file or directory'

#> Error in gzfile(file, "rb"): cannot open the connection

df = df[df$date < as.Date("2021-01-01"), ]

#> Error in df$date: object of type 'closure' is not subsettable

day_diff = as.numeric(df$date[nrow(df)] - df$date[1])

#> Error in df$date: object of type 'closure' is not subsettable

v = df$count

#> Error in df$count: object of type 'closure' is not subsettable

v[v > quantile(v, 0.95)] = quantile(v, 0.95)

#> Error in quantile(v, 0.95): object 'v' not found

spiral_initialize(xlim = c(0, nrow(df)), start = 360, end = 360*(day_diff/364)) # a circle 52 weeks

#> Error: Length of `xlim` should be two.

spiral_track(height = 0.9)

#> Error in data.frame(i = track_index, ymin = ylim[1], ymax = ylim[2], rmin = sum_height *     dist, rmax = (sum_height + height) * dist, rel_height = height): arguments imply differing number of rows: 1, 0

spiral_horizon(1:nrow(df), v)

#> Error in if (!(get_track_data("ymin", track_index) == 0 & get_track_data("ymax",     track_index) == 1)) {    stop_wrap("The horizon track must have 'ylim = c(0, 1)'.")}: argument is of length zero

spiral_initialize(xlim = c(0, nrow(df)), start = 360, end = 360*(day_diff/364))

#> Error: Length of `xlim` should be two.

spiral_track(height = 0.9)

#> Error in data.frame(i = track_index, ymin = ylim[1], ymax = ylim[2], rmin = sum_height *     dist, rmax = (sum_height + height) * dist, rel_height = height): arguments imply differing number of rows: 1, 0

spiral_horizon(1:nrow(df), v, use_bars = TRUE)

#> Error in if (!(get_track_data("ymin", track_index) == 0 & get_track_data("ymax",     track_index) == 1)) {    stop_wrap("The horizon track must have 'ylim = c(0, 1)'.")}: argument is of length zero

# }